anon patch, s/cos/cosf/, same for sin
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@8784 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/WorldWeapons.cxx
@@ -85,8 +85,8 @@ void WorldWeapons::fire()
       firingInfo.shot.player = ServerPlayer;
       memmove(firingInfo.shot.pos, w->origin, 3 * sizeof(float));
       float shotSpeed = BZDB.eval(StateDatabase::BZDB_SHOTSPEED);
-      firingInfo.shot.vel[0] = shotSpeed*cos(w->direction);
-      firingInfo.shot.vel[1] = shotSpeed*sin(w->direction);
+      firingInfo.shot.vel[0] = shotSpeed * cosf(w->direction);
+      firingInfo.shot.vel[1] = shotSpeed * sinf(w->direction);
       firingInfo.shot.vel[2] = 0.0f;
       firingInfo.shot.id = worldShotId++;
       if (worldShotId > 30) { // Maximum of 30 world shots
